Catch rare exceptions from fs::remove instead of crashing.
On one of our CI machines, NRD's build would sometimes fail at random
points while generating embedded shader headers using ShaderMake.
This would result in a cryptic error message like this:

```
Compiling shaders using: C:/VulkanSDK/1.4.313.2/Bin/dxc.exe
  ( 10%) NRD SPIRV
  The system cannot find the batch label specified - VCEnd
```

The `VCEnd` message is a CMake bug: see https://gitlab.kitware.com/cmake/cmake/-/issues/26678 .

It turned out that on that machine, a handle to the output file would
sometimes still be held by a `cmd` subprocess. This should never
happen since we call `pclose`, which waits for the subprocess to
terminate, but it seems that on this specific system it sometimes does.

This patch catches the exception if it occurs and instead prints
a warning message that looks like this:

```
  Compiling shaders using: C:/VulkanSDK/1.4.313.2/Bin/dxc.exe
  ( 10%) NRD SPIRV
  Could not delete temporary binary file: 'remov
  e: The process cannot access the file because it is being used by another process.: "C:\Users\nbickford\Documents\NRD
  \_Shaders\SIGMA_SmoothTiles.cs.spirv"'
  ( 20%) NRD SPIRV
  ( 30%) NRD SPIRV
  ( 40%) NRD SPIRV
  ( 50%) NRD SPIRV
  ( 60%) NRD SPIRV
  ( 70%) NRD SPIRV
  ( 80%) NRD SPIRV
  ( 90%) NRD SPIRV
  (100%) NRD SPIRV
  159 task(s) completed successfully (elapsed time 5663.02 ms)
```